Put all validation rules in a file
